Search functionality with filters Added more fields in listing Changed Ids from Integer to Long Most of the app is functional now

+++ src/main/java/com/schoolproject/airbnbclone/repositories/ImageRepository.java
@@ -11,11 +11,11 @@ public interface ImageRepository extends JpaRepository<Image, Integer> {
     @Transactional
     @Modifying
     @Query("DELETE FROM Image i WHERE i.listing.id = ?1")
-    void deleteByListing(Integer listingID);
+    void deleteByListing(Long listingID);
 
     @Transactional
     @Modifying
     @Query("DELETE FROM Image i WHERE i.user.id = ?1")
-    void deleteByUser(Integer userID);
+    void deleteByUser(Long userID);
 
 }
